+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ## Changed
 * `addCommerceProductImpression()` no longer supports list index and list name. ([#6](https://github.com/nystudio107/craft-instantanalytics-ga4/issues/6))
 
+## Fixed
+* Fixed Craft 3.6.x compatibility issue. ([#11](https://github.com/nystudio107/craft-instantanalytics-ga4/issues/11))
+
 ## 3.0.0-beta.1 - 2023.06.25
 ### Added
 * Initial beta release